添加了注册和通知界面的ui文件
