+++ build.bat
@@ -2,30 +2,18 @@
 
 setlocal EnableDelayedExpansion
 
-set "TOTAL=0"
-set "FAIL=0"
+set "logDate=%date:~0,4%-%date:~5,2%-%date:~8,2% %time:~0,2%:%time:~3,2%:%time:~6,2%"
 
-start "format" /wait cmd /c ".\script\format.bat"
-set /a TOTAL+=1
-if errorlevel 1 set /a FAIL+=1
+call ".\script\format.bat"
+if errorlevel 1 echo !logDate! [FAIL] format>>"log\BuildFail.log" & echo. >>"log\BuildFail.log"
 
-start "build_Debug" /wait cmd /c ".\script\build_Debug.bat"
-set /a TOTAL+=1
-if errorlevel 1 set /a FAIL+=1
+call ".\script\build_Debug.bat"
+if errorlevel 1 echo !logDate! [FAIL] debug>>"log\BuildFail.log" & echo. >>"log\BuildFail.log"
 
-start "build_Release" /wait cmd /c ".\script\build_Release.bat"
-set /a TOTAL+=1
-if errorlevel 1 set /a FAIL+=1
+call ".\script\build_Release.bat"
+if errorlevel 1 echo !logDate! [FAIL] release>>"log\BuildFail.log" & echo. >>"log\BuildFail.log"
 
-start "doc" /wait cmd /c ".\script\doc.bat"
-set /a TOTAL+=1
-if errorlevel 1 set /a FAIL+=1
-
-echo ===== Summary =====
-echo Total : %TOTAL%
-echo Failed: %FAIL%
-echo ===================
-
-pause
+call ".\script\doc.bat"
+if errorlevel 1 echo !logDate! [FAIL] doc>>"log\BuildFail.log" & echo. >>"log\BuildFail.log"
 
 cls
