+++ include/View/Chat/NotificationBadge.h
@@ -1,10 +1,9 @@
-// NotificationBadge.h
+/* guard ------------------------------------------------------------------ 80 // ! ----------------------------- 120 */
 
 #ifndef INCLUDE_VIEW_CHAT_NOTIFICATION_BADGE
 
 #define INCLUDE_VIEW_CHAT_NOTIFICATION_BADGE
 
-/* guard ------------------------------------------------------------------ 80 // ! ----------------------------- 120 */
 /* include ---------------------------------------------------------------- 80 // ! ----------------------------- 120 */
 
 #include <QWidget>
