+++ include/View/Friend/SearchFriendWindow.h
@@ -7,8 +7,7 @@
 /* include ---------------------------------------------------------------- 80 // ! ----------------------------- 120 */
 
 #include "Components/LineEditComponent.h"
-#include "Utils/FramelessWindow.h"
-
+#include "Util/FramelessWindow.h"
 #include "View/Friend/SearchResultList.h"
 #include "View/Friend/SearchTypeTab.h"
