+++ script/doc.bat
@@ -1,7 +1,9 @@
 @echo off
 
+cd /d "%~dp0.."
+
 doxygen Doxyfile
 
 copy /y ".\doc\html\NetherLink-Client.chm" ".\doc\NetherLink-Client.chm"
 
-timeout /t 10
+cls
