+++ src/Network/NetworkConfig.cpp
@@ -2,7 +2,6 @@
 
 #include <QCoreApplication>
 #include <QFile>
-#include <QJsonDocument>
 #include <QJsonObject>
 
 #include "Network/NetworkConfig.h"
