+++ src/Util/Validator.cpp
@@ -1,0 +1,31 @@
+/**
+ * @file Validator.cpp
+ * @version 1.0.0
+ * @author 落羽行歌 (2481036245@qq.com)
+ * @date 2025-08-14 周四 09:32:03
+ * @brief 【描述】
+ */
+
+/* include ---------------------------------------------------------------- 80 // ! ----------------------------- 120 */
+
+#include "Util/Validator.h"
+
+/* namespace -------------------------------------------------------------- 80 // ! ----------------------------- 120 */
+namespace Util {
+/* variable --------------------------------------------------------------- 80 // ! ----------------------------- 120 */
+    const QRegularExpression Validator::邮箱正则表达式("^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$");
+
+/* function --------------------------------------------------------------- 80 // ! ----------------------------- 120 */
+
+    /**
+     * @brief 【描述】 检查邮箱合法性
+     *
+     * @param email 【参数注释】 邮箱号
+     *
+     * @return  true 【返回值注释】 邮箱号合法
+     * @return  false 【返回值注释】 邮箱号非法
+     */
+    bool Validator::函数_是否为有效邮箱(const QString& 邮箱) {
+        return (邮箱正则表达式.match(邮箱).hasMatch());
+    }
+}
